backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
Werkzeug==2.3.7
stripe==5.4.0
sendgrid==6.10.0
python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
scikit-learn==1.3.1
xgboost==2.0.0
lightgbm==4.0.0
catboost==1.2.2
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
